:fire:fix

+++ generate.go
@@ -7,10 +7,12 @@ type PressureTestResult struct {
 	duration_time  float32
 }
 
-func (r PressureTestResult) success_rate() int {
-	return (r.request_number - r.failure_number) / r.request_number * 100
+// get Success rate
+func (r PressureTestResult) success_rate() float32 {
+	return float32(r.request_number - r.failure_number) / float32(r.request_number) * 100
 }
 
+// get Average time consuming
 func (r PressureTestResult) average_time() float32 {
 	return float32(r.request_number) / r.duration_time
 }
